Registers: Move in register test from `info.rs`:
Removes the need to make the inner hash map tuple field public.
