+++ helix-view/src/register.rs
@@ -5,7 +5,7 @@ pub type Register = Vec<String>;
 
 /// Currently just wraps a `HashMap` of `Register`s
 #[derive(Debug, Clone, Default, Deref, DerefMut)]
-pub struct Registers(pub HashMap<char, Register>);
+pub struct Registers(HashMap<char, Register>);
 
 impl Registers {
     pub fn read(&self, name: char) -> Option<&[String]> {
@@ -34,5 +34,27 @@ impl Registers {
 
     pub fn last(&self, name: char) -> Option<&String> {
         self.read(name).and_then(|entries| entries.last())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::info::Info;
+    use once_cell::sync::Lazy;
+
+    const REGISTER_VALUE_1_MOCK: &str = "value_1";
+    const REGISTER_VALUE_2_MOCK: &str = "value_2";
+    static REGISTERS_MOCK: Lazy<Registers> =
+        Lazy::new(|| Registers([('/', vec![REGISTER_VALUE_1_MOCK.to_string()])].into()));
+
+    #[test]
+    fn infobox_shows_latest_value() {
+        let mut registers = (*REGISTERS_MOCK).clone();
+        registers.push('/', REGISTER_VALUE_2_MOCK.to_string());
+
+        assert!(Info::from_registers(&registers)
+            .text
+            .contains(REGISTER_VALUE_2_MOCK));
     }
 }
